框架结构调整，控制器初始化方法修改:initialize->init,Trce 类调整

+++ DolrPHP/Dao/Adapter.php
@@ -19,8 +19,6 @@ abstract class DB_Adapter
     /**
      * FETCH TYPE
      */
-    const FETCH_ALL         = 'all';
-    const FETCH_ONE         = 'one';
     const FETCH_TYPE_ARRAY  = 'array';
     const FETCH_TYPE_NUM    = 'num';
     const FETCH_TYPE_ASSOC  = 'assoc';
@@ -121,7 +119,7 @@ abstract class DB_Adapter
     {
         $tableMeta = $this->_getTableMetaInfo($tableName);
         if (!$tableMeta || empty($tableMeta)) {
-            throw new Exception("数据表 '{$tableName}' 读取失败", 1);
+            throw new Exception("数据表 '{$tableName}' 不存在或读取失败", 1);
         }
         $this->_tableMeta = $tableMeta;
 
@@ -137,7 +135,7 @@ abstract class DB_Adapter
      */
     protected function _getTableMetaInfo($tableName)
     {
-        $tableInfo  = $this->query("SHOW COLUMNS FROM `$tableName`");
+        $tableInfo = $this->query("SHOW COLUMNS FROM `$tableName`");
         $data = array();
         $data['_name']   = $tableName;
         $data['_fields'] = array();
@@ -165,14 +163,13 @@ abstract class DB_Adapter
      * 过滤输入数据
      *
      * @param array  $data   input data
-     * @param string $action insert|update
      *
      * @return array
      */
-    protected function _filterFields($data, $action = 'insert')
+    protected function _filterFields($data)
     {
         if (empty($this->_tableMeta)) {
-            throw new Exception("未选择目标数据表[{$action}]", 1);
+            throw new Exception("未初始化目标数据表");
         }
         //清空values
         $output = array();
@@ -189,14 +186,16 @@ abstract class DB_Adapter
     /**
      * 执行一个完整SQL查询
      *
-     * @param string $sql    SQL
-     * @param array  $params values to bind
+     * @param string $sql       SQL
+     * @param array  $params    values to bind
+     * @param array  $fetchType fetch type (assoc|num|array|object)
      *
      * @return mixed
      */
-    public function query($sql, array $params = array(), $fetch = self::FETCH_ALL, $fetchType = self::FETCH_TYPE_ASSOC)
+    public function query($sql, array $params = array(), $fetchType = self::FETCH_TYPE_ASSOC)
     {
         $sqlType = preg_match('/([a-z]+)\s+/i', $sql, $matches);
+        $sql = rtrim($sql, ';');
         switch (strtoupper($matches[1])) {
             case 'INSERT':
                 $this->_connector = &$this->_writer;
@@ -212,10 +211,11 @@ abstract class DB_Adapter
             default:
                 $this->_connector = &$this->_reader;
                 $res = $this->exec($sql, $params);
-                $ret = $this->fetch($res, $fetch, $fetchType);
+                $ret = $this->_fetch($res, $fetchType);
                 break;
         }
         $this->_setLastSql($sql, $params);
+
         return $ret;
     }
 
@@ -246,12 +246,11 @@ abstract class DB_Adapter
      * 提取结果集
      *
      * @param resource $resource  query resource
-     * @param string   $fetch     one | all
      * @param string   $fetchType fetch type [array, num, assoc, object]
      *
      * @return array or boolean
      */
-    public function fetch($resource, $fetch, $fetchType)
+    protected function _fetch($resource, $fetchType)
     {
         if (!is_resource($resource) && !is_object($resource)) {
             return $resource;
@@ -273,13 +272,25 @@ abstract class DB_Adapter
                 $res = false;
                 break;
         }
-        if (!$res) {
+        if (false === $res) {
             return false;
         }
-        if ($fetch == self::FETCH_ONE) {
-            return array_shift($res);
-        }
+
         return $res;
+    }
+
+    /**
+     * 提取数组中的第一个元素
+     *
+     * @param array $array 结果数组
+     * @return mixed
+     */
+    public function fetchOne($array)
+    {
+        if (empty($array)) {
+            return $array;
+        }
+        return array_shift($array);
     }
 
     /**
@@ -295,8 +306,11 @@ abstract class DB_Adapter
      *
      * @return int
      */
-    public function add($data)
+    public function add(array $data)
     {
+        if (empty($data)) {
+            return false;
+        }
         $sql = $this->_createSql(self::SQL_TYPE_INSERT, '', $data);
         $res = $this->query($sql,$data);
         if (!$res) {
@@ -316,7 +330,7 @@ abstract class DB_Adapter
     public function del($sql, array $values = array())
     {
         $sql = $this->_createSql(self::SQL_TYPE_DELETE, $sql, $values);
-        return $this->query($sql,$values);
+        return $this->query($sql, $values);
     }
 
     /**
@@ -324,13 +338,15 @@ abstract class DB_Adapter
      *
      * @param string $sql    SQL
      * @param array  $values values to bind
+     * @param array  $fetchType fetch type (assoc|num|array|object)
      *
      * @return array
      */
-    public function find($sql = '', array $values = array())
+    public function find($sql = '', array $values = array(), $fetchStyle = self::FETCH_TYPE_ASSOC)
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql);
-        return $this->query($sql,$values, self::FETCH_ONE);
+        $sql .= " LIMIT 1";
+        return $this->_fetchOne($this->query($sql, $values, $fetchStyle));
     }
 
     /**
@@ -338,27 +354,30 @@ abstract class DB_Adapter
      *
      * @param string $sql    SQL
      * @param array  $values values to bind
-     *
+     * @param array  $fetchType fetch type (assoc|num|array|object)
      * @return array
      */
-    public function select($sql = '', array $values = array())
+    public function select($sql = '', array $values = array(), $fetchStyle = self::FETCH_TYPE_ASSOC)
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql);
-        return $this->query($sql,$values, self::FETCH_ALL);
+        return $this->query($sql, $values, $fetchStyle);
     }
 
 
     /**
      * 更新记录
      *
-     * @param array  $data
-     * @param string $sql
-     * @param array  $values
+     * @param array  $data   data to insert
+     * @param string $sql    where condition
+     * @param array  $values values to bind
      *
      * @return bool|mixed
      */
     public function save(array $data, $sql = '')
     {
+        if (empty($data)) {
+            return false;
+        }
         $data = $this->_filterFields($data);
         $sql = $this->_createSql(self::SQL_TYPE_UPDATE, $sql, $data);
         $res = $this->query($sql,$data);
@@ -367,6 +386,38 @@ abstract class DB_Adapter
         }
 
         return $res;
+    }
+
+    /**
+     * 设置一条记录
+     * 如果不存在则add($data),存在则save($data)
+     *
+     * @param array $data 数据
+     * @param string $checkFields 需要检测的字段，判断存在与否的字段
+     */
+    public function set(array $data, $checkFields)
+    {
+        if (empty($data)) {
+            return false;
+        }
+        $data = $this->_filterFields($data);
+        if (empty($checkFields)) {
+            $checkFields = array_keys($data);
+        } elseif (is_string($checkFields)) {
+            $checkFields = stripos($checkFields, ',') ?
+                            explode(',', $checkFields) : array(trim($checkFields));
+        }
+        $condition = array();
+        foreach ($checkFields as $key) {
+            if (!isset($data[$key])) {
+                continue;
+            }
+            $condition = "`{$key}` = ?";
+        }
+        $sql = "WHERE ".join(' AND ', $condition);
+        if ($this->find($sql, $data)) {
+            $this->save($data);
+        }
     }
 
     /**
@@ -380,7 +431,8 @@ abstract class DB_Adapter
     public function getRow($sql = '', array $values = array())
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        return $this->query($sql, $values, self::FETCH_ONE);
+        $sql .= " LIMIT 1";
+        return $this->_fetchOne($this->query($sql, $values));
     }
 
     /**
@@ -390,10 +442,10 @@ abstract class DB_Adapter
      *
      * @return array
      */
-    public function getAll($sql = '', array $values = array())
+    public function getAll($sql = '', array $values = array(), $fetchStyle = self::FETCH_TYPE_ASSOC)
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        return $this->query($sql, $values, self::FETCH_ALL);
+        return $this->query($sql, $values, $fetchStyle);
     }
 
     /**
@@ -422,20 +474,25 @@ abstract class DB_Adapter
      */
     public function getCol($colName, $convertTo2D = false, $sql = '', array $values = array())
     {
+        if (empty($this->_tableMeta)) {
+            throw new Exception("未初始化目标数据表");
+        }
         if (!in_array($colName, $this->_tableMeta['_fields'])) {
             $this->_log("表'{$this->_tableMeta['_name']}'不存在字段'{$colName}'");
             return false;
         }
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        $result = $this->query($sql, $values, self::FETCH_ALL);
+        $result = $this->query($sql, $values);
         $cols = array();
+        $cols2D = array();
         foreach ($result as $value) {
             if (isset($value[$colName])) {
                 $cols[][$colName] = $value[$colName];
+                $cols2D[] = $value[$colName];
             }
         }
 
-        return $cols;
+        return $convertTo2D ? $cols2D : $cols;
     }
 
     /**
@@ -451,8 +508,13 @@ abstract class DB_Adapter
      */
     public function getCell($cellName, $sql = '', array $values = array())
     {
-        $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        $res = $this->query($sql, $values, self::FETCH_ONE);
+        if (!in_array($colName, $this->_tableMeta['_fields'])) {
+            $this->_log("表'{$this->_tableMeta['_name']}'不存在字段'{$cellName}'");
+            return false;
+        }
+        $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values, "`$cellName`");
+        $sql .= "LIMIT 1";
+        $res = $this->_fetchOne($this->query($sql, $values));
         if ($res && isset($res[$cellName])) {
             return $res[$cellName];
         }
@@ -472,7 +534,8 @@ abstract class DB_Adapter
     public function getAssoc($sql = '', array $values = array())
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        return $this->query($sql, $values, self::FETCH_ONE, self::FETCH_TYPE_ASSOC);
+        $sql .= "LIMIT 1";
+        return $this->_fetchOne($this->query($sql, $values, self::FETCH_TYPE_ASSOC));
     }
 
     /**
@@ -487,7 +550,8 @@ abstract class DB_Adapter
     public function getObject($sql = '', array $values = array())
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        return $this->query($sql, $values, self::FETCH_ONE, self::FETCH_TYPE_OBJECT);
+        $sql .= "LIMIT 1";
+        return $this->_fetchOne($this->query($sql, $values, self::FETCH_TYPE_OBJECT));
     }
 
     /**
@@ -502,7 +566,7 @@ abstract class DB_Adapter
     public function getObjects($sql = '', array $values = array())
     {
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values);
-        return $this->query($sql, $values, self::FETCH_ALL, self::FETCH_TYPE_OBJECT);
+        return $this->query($sql, $values, self::FETCH_TYPE_OBJECT);
     }
 
     /**
@@ -516,13 +580,16 @@ abstract class DB_Adapter
      */
     public function getCount($field = '', $sql = '', array $values = array())
     {
+        if (empty($this->_tableMeta)) {
+            throw new Exception("未初始化目标数据表");
+        }
         if (empty($field) || !is_array($field, $this->_tableMeta['_fields'])) {
             $field = $this->_tableMeta['_pk'];
         }
         $sql = $this->_createSql(self::SQL_TYPE_SELECT, $sql, $values, "COUNT(`{$field}`) AS `count`");
-        $res = $this->query($sql, $values);
+        $sql .= "LIMIT 1";
+        $res = $this->_fetchOne($this->query($sql, $values));
         if ($res) {
-            $res = array_shift($res);
             return $res['count'];
         }
     }
@@ -538,29 +605,40 @@ abstract class DB_Adapter
      */
     protected function _createSql($sqlType, $sql, $data = array(), $fieldArea = '*')
     {
+        if (empty($this->_tableMeta)) {
+            throw new Exception("未初始化目标数据表");
+        }
         if (!empty($data)) {
             $data = $this->_filterFields($data);
             $fields = array_keys($data);
+        } else {
+            $fields = $this->tableMeta['fields'];
         }
         switch ($sqlType) {
             case self::SQL_TYPE_SELECT:
-                $sql = "SELECT {$fieldArea} FROM `{$this->_tableMeta['_name']}` {$sql}";
+                $sql = "SELECT {$fieldArea} FROM `{$this->_tableMeta['_name']}` {$sql} ";
                 break;
             case self::SQL_TYPE_DELETE:
-                $sql = "DELETE FROM `{$this->_tableMeta['_name']}` {$sql}";
+                $sql = "DELETE FROM `{$this->_tableMeta['_name']}` {$sql} ";
                 break;
             case self::SQL_TYPE_UPDATE:
+                if (empty($data)) {
+                    return false;
+                }
                 $arr = array();
                 foreach ($fields as $field) {
                     $arr[] = "`{$field}` = ? ";
                 }
                 $setString = join(',', $arr);
-                $sql = "UPDATE `{$this->_tableMeta['_name']}` SET {$setString} {$sql}";
+                $sql = "UPDATE `{$this->_tableMeta['_name']}` SET {$setString} {$sql} ";
                 break;
             case self::SQL_TYPE_INSERT:
+                if (empty($data)) {
+                    return false;
+                }
                 $keys = join(',', $fields);
                 $valuesFlag = join(',', array_fill(0, count($fields), '?'));
-                $sql = "INSERT INTO `{$this->_tableMeta['_name']}`({$keys}) VALUES({$valuesFlag})";
+                $sql = "INSERT INTO `{$this->_tableMeta['_name']}`({$keys}) VALUES({$valuesFlag}) ";
                 break;
             default:
                 break;
@@ -577,9 +655,12 @@ abstract class DB_Adapter
      */
     public function __get($proName)
     {
-        switch (strtoupper($proName)) {
-            case 'LASTSQL':
+        switch (strtolower($proName)) {
+            case 'lastsql':
                 return $this->_lastSql;
+                break;
+            case 'insertid':
+                return $this->_lastInsertId;
                 break;
             default:
                 # code...
